Clean up, tests run locally so not sure why failing
