+++ src/DirectNumericalCabbelingShenanigans.jl
@@ -9,13 +9,7 @@ import Base: show
 
 @reexport using Oceananigans, Reexport
 
-export
-    TwoLayerDNS,
-    SIMULATION_PATH,
-    DNCS,
-    DNS,
-    DNS_simulation_setup,
-    non_dimensional_numbers
+export TwoLayerDNS, DNS, DNS_simulation_setup
 
 export
     StableUpperLayerInitialConditions,
@@ -37,21 +31,17 @@ export ContinuousProfileFunction, HyperbolicTangent, Erf
 export SalinityPerturbation, GaussianProfile, GaussianBlob, RandomPerturbations
 
 export DOMAIN_EXTENT, HIGH_RESOLUTION, SO_DIFFUSIVITIES, REFERENCE_DENSITY,
-       INTERFACE_LOCATION
+       INTERFACE_LOCATION, SIMULATION_PATH, DNCS
 
-export compute_density, compute_density!
+export compute_density, compute_density!, non_dimensional_numbers
 
-export
-    animate_2D_field,
-    visualise_initial_conditions,
-    visualise_initial_density,
-    visualise_snapshot
+export animate_2D_field, visualise_initial_conditions, visualise_initial_density,
+       visualise_snapshot
 
 include("initialconditions.jl")
 include("continuousprofilefunctions.jl")
 include("salinityperturbations.jl")
 include("twolayerdns.jl")
-include("dns.jl")
 include("set_initialconditions.jl")
 include("makiefunctions.jl")
 include("constants.jl")
